+++ server/odc-service/src/main/java/com/oceanbase/odc/metadb/connection/DatabaseEntity.java
@@ -30,7 +30,7 @@ import org.hibernate.annotations.Generated;
 import org.hibernate.annotations.GenerationTime;
 
 import com.oceanbase.odc.service.connection.database.model.DatabaseSyncStatus;
-import com.oceanbase.odc.service.db.model.DBObjectSyncStatus;
+import com.oceanbase.odc.service.db.object.model.DBObjectSyncStatus;
 
 import lombok.Data;
