+++ server/odc-service/src/main/java/com/oceanbase/odc/service/db/object/model/DBObjectSyncStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.db.model;
+package com.oceanbase.odc.service.db.object.model;
 
 /**
  * @author gaoda.xy
